MDEV-18202: SQL_CALC_FOUND_ROWS do not work with UNION ALL
select_options (here used in conjunction with OPTION_FOUND_ROWS) is not set
correctly to check if the SELECT in union of SELECTs has SQL_CALC_FOUND_ROWS.
As a result, (select_options & OPTION_FOUND_ROWS) would give 0 and
send_records gets assigned 0 (number of rows sent to client in the query.
In case of LIMIT 0, no rows were sent to client, so send_records gets
assigned 0).

Since thd->limit_found_rows(the variable from where FOUND_ROWS() gets
its value from) gets its value from send_records, it also gets assigned 0.

Fix: The fix is made so that it is consistent with the tabled queries.
If we don’t have SQL_CALC_FOUND_ROWS in SELECT then we drop the flag.
If we have it in SELECT, we need to check for braces and LIMIT.
When there are no braces we keep the flag because we want the total number
of rows that are actually there in the table and not the number of rows
that were sent to the client (because even if we have LIMIT in this case,
it will be global LIMIT instead of local LIMIT). If we have braces
and LIMIT for a particular select, we don't want the total number of rows
that could have been sent to client if LIMIT was not present,
but we want the number of rows that were actually sent to client
(value returned by thd->get_sent_row_count()), irrespective if we have
SQL_CALC_FOUND_ROWS. Now send_records gets the value returned by
thd->get_sent_row_count() (number of rows that were sent to client)
if the flag is off, so we need to drop the flag. If we have braces
but no LIMIT, we need the number of rows that were sent to
client (1 for tabless queries). Now send_records is 1 when we have the
flag, so we don't drop the flag.

After setting the flag correctly, thd->limit_found_rows=send_records
needs to be done before send_eof() because send_eof() is incrementing
limit_found_rows and also assigning thd->limit_found_row = limit_found_rows
when current_select==last_select_lex.
Keeping thd->limit_found_rows = send_record after send_eof() will
reset the value of thd->limit_found_rows to 1 (because send_records is 1
for each union of select in case of tables queries) even when it
gets incremented in send_eof().