MDEV-21201 No records produced in information_schema query, depending on projection
In case of USING check also join field list to mark fields as used.

In case of NATURAL JOIN mark all field (one table can not be opened
in any case so optimisation does not worth it).

IMHO table should be checked for used fields and filled after prepare,
when we will fave whole info about used fields but it is too big change
for a bugfix.