MDEV-23449: alias do not exist and a query do not report an error
The issue here was for an IN subquery without an aggregate function, the GROUP BY clause
was removed. Due to the GROUP BY list being removed, the invalid reference in the GROUP BY
clause was never resolved.
Remove the GROUP BY listonly when the all the items in the
GROUP BY list are resolved.