MDEV-23160: SIGSEGV in Explain_node::print_explain_for_children on UNION SELECT or on EXPLAIN EXTENDED
The issue here was that the ORDER BY clause had a subquery and the
ORDER BY clause was defined for a UNION inside an IN subquery.
For IN/ALL/ANY subquery the ORDER BY clause makes no sense and should be removed.
But the removal of the ORDER BY clause happened before name resolution. So an invalid query
could also be parsed and would not return an error.

The reason for the crash is that for EXPLAIN of UNION, the fake_select_lex adds the
removed subquery as its child. The subquery being removed does not create a node for itself
for EXPLAIN. But when the EXPLAIN is printed for the fake_select_lex it tries to print the nodes
of its children.

The fix would be to mark the subquery to be eliminated so that it is not added as a child
to the fake_select_lex node. Then after name resolution is done and the query is valid
then remove the subquery so that it does not get executed.