MDEV-23633 UT_RELAX_CPU or MY_RELAX_CPU does unnecessary compare-and-swap
This is a follow-up to MDEV-14374. On target architectures where nothing
special is available (not IA-32, AMD64, POWER), we would perform a
dummy compare-and-swap operation. This is contrary to the idea of the
x86 PAUSE instruction and the __ppc_get_timebase(), which aim to keep
the memory bus idle for a while, to allow other cores to better execute
code while a spinloop is waiting for something to be changed.