MDEV-23535: SIGSEGV, SIGABRT and SIGILL in typeinfo for Item_func_set_collation (on optimized builds)
Issue:

For single row subselects, the result is stored in Item_cache.
For eg SELECT (SELECT a, b FROM t1) FROM t2 , we wrap the column a and b inside an Item_cache.
For the first iteration we would calculate the value for the column a and b and store the
value in the cache and for subsequent iteration we can just read the cached value.

But in the query the issue happens that the comparison of expression involve expressions
with different character sets, we create an Item_func_conv_charset, to do the conversion.
The implementation for single row subselects expects the columns to be wrapped
inside Item_cache instance but this is not the case here.

Fix:
Repleace Item_cache** with Item**, so that we can also store the items if Item_cache gets
converted to a different item. Virtual functions are added so that these virtual function
are called and this would in turn call the Item_cache specific functions.